Further refactored cloud test code
