+++ tests/integration/cloud/clouds/test_gce.py
@@ -8,7 +8,7 @@
 from __future__ import absolute_import, print_function, unicode_literals
 
 # Import Salt Testing Libs
-from tests.integration.cloud.cloud_test_helpers import TIMEOUT, CloudTest
+from tests.integration.cloud.helpers.cloud_test_base import TIMEOUT, CloudTest
 from tests.support.paths import FILES
 from tests.support.helpers import expensiveTest
 
@@ -77,6 +77,7 @@ class GCETest(CloudTest):
         # check if instance returned with salt installed
         self.assertIn(ret_str, instance)
         self.assertEqual(self._instance_exists(), True)
+        self._destroy_instance()
 
     def test_instance_extra(self):
         '''
@@ -92,3 +93,4 @@ class GCETest(CloudTest):
         # check if instance returned with salt installed
         self.assertIn(ret_str, instance)
         self.assertEqual(self._instance_exists(), True)
+        self._destroy_instance()
