+++ tests/integration/cloud/clouds/test_oneandone.py
@@ -23,7 +23,6 @@ except ImportError:
 INSTANCE_NAME = generate_random_name('CLOUD-TEST-')
 PROVIDER_NAME = 'oneandone'
 DRIVER_NAME = 'oneandone'
-TIMEOUT = 500
 
 
 @skipIf(HAS_ONEANDONE is False, 'salt-cloud requires >= 1and1 1.2.0')
@@ -71,9 +70,6 @@ class OneAndOneTest(CloudTest):
         self.assertEqual(self._instance_exists(), False,
                          'The instance "{}" exists before it was created by the test'.format(INSTANCE_NAME))
 
-    def _instance_exists(self):
-        return '        {0}:'.format(INSTANCE_NAME) in self.run_cloud('--query')
-
     def test_list_images(self):
         '''
         Tests the return of running the --list-images command for 1and1
@@ -89,26 +85,12 @@ class OneAndOneTest(CloudTest):
         Test creating an instance on 1and1
         '''
         # check if instance with salt installed returned
-        try:
-            self.assertIn(
-                INSTANCE_NAME,
-                [i.strip() for i in self.run_cloud(
-                    '-p oneandone-test {0}'.format(INSTANCE_NAME), timeout=500
-                )]
-            )
-        except AssertionError:
-            self.run_cloud('-d {0} --assume-yes'.format(INSTANCE_NAME), timeout=500)
-            raise
+        self.assertIn(
+            INSTANCE_NAME,
+            [i.strip() for i in self.run_cloud(
+                '-p oneandone-test {0}'.format(INSTANCE_NAME), timeout=TIMEOUT
+            )]
+        )
+        self.assertEqual(self._instance_exists(), True)
 
-    def tearDown(self):
-        '''
-        Clean up after tests
-        '''
-        if self._instance_exists():
-            delete = self.run_cloud('-d {0} --assume-yes'.format(INSTANCE_NAME), timeout=TIMEOUT)
-            # example response: ['gce-config:', '----------', '    gce:', '----------', 'cloud-test-dq4e6c:', 'True', '']
-            delete_str = ''.join(delete)
-
-            # check if deletion was performed appropriately
-            self.assertIn(INSTANCE_NAME, delete_str)
-            self.assertIn('True', delete_str)
+        self._destroy_instance()
